use confirm hook

+++ components/data-table.tsx
@@ -25,6 +25,7 @@ import { Button } from "@/components/ui/button"
 import React from "react"
 import { Input } from "@/components/ui/input"
 import { Trash } from "lucide-react"
+import { useConfirm } from "@/hooks/use-confirm"
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[]
@@ -41,6 +42,11 @@ export function DataTable<TData, TValue>({
   onDelete,
   disable,
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmationDialog, confirm] = useConfirm(
+    "Deletar",
+    "Tem certeza que deseja deletar as linhas selecionadas?"
+  )
+
   const [sorting, setSorting] = React.useState<SortingState>([])
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
@@ -66,6 +72,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmationDialog />
       <div className="flex items-center py-4">
         <Input
           placeholder={`Filtrar por ${filterKey} ...`}
@@ -81,9 +88,12 @@ export function DataTable<TData, TValue>({
             size={"sm"}
             variant={"outline"}
             className="ml-auto font-normal text-xs"
-            onClick={() => {
-              onDelete(table.getFilteredSelectedRowModel().rows)
-              table.resetRowSelection()
+            onClick={async () => {
+              const ok = await confirm();
+              if (ok) {
+                onDelete(table.getFilteredSelectedRowModel().rows)
+                table.resetRowSelection()
+              }
             }}
           >
             <Trash className="size-4 mr-2" />
